+++ frontend/openchat-agent/src/constants.ts
@@ -4,3 +4,4 @@ export const MAX_MISSING = 30;
 export const OPENCHAT_BOT_USER_ID = "zzyk3-openc-hatbo-tq7my-cai";
 export const OPENCHAT_BOT_USERNAME = "OpenChatBot";
 export const OPENCHAT_BOT_AVATAR_URL = "assets/robot.svg";
+export const ONLINE_THRESHOLD = 120;
