+++ frontend/openchat-agent/src/index.ts
@@ -10,4 +10,17 @@ export {
     indexRangeForChat,
     userIdsFromEvents,
     getDisplayDate,
+    eventIsVisible,
 } from "./utils/chat";
+
+export { getUserStatus, userStatus, missingUserIds } from "./domain/user/user.utils";
+
+export { UnsupportedValueError } from "./utils/error";
+
+export { getTimeUntilSessionExpiryMs } from "./utils/session";
+export {
+    setCachedMessageFromNotification,
+    getSoftDisabled,
+    storeSoftDisabled,
+} from "./utils/caching";
+export { Logger } from "./utils/logging";
