+++ frontend/openchat-client/src/utils/chat.spec.ts
@@ -1,62 +1,24 @@
-import type {
-    DirectChatSummary,
+import {
     GroupChatSummary,
-    DirectChatSummaryUpdates,
-    GroupChatSummaryUpdates,
-    UpdatesResponse,
     PollConfig,
     PollVotes,
     PollContent,
     PartialUserSummary,
     UserLookup,
     UserSummary,
+    emptyChatMetrics,
 } from "openchat-agent";
 import {
     addVoteToPoll,
-    enoughVisibleMessages,
     getMembersString,
     mergeChatMetrics,
-    mergeChatUpdates,
     mergeUnconfirmedThreadsIntoSummary,
-    newMessageId,
-} from "./chat.utils";
-import { emptyChatMetrics } from "./chat.utils.shared";
+} from "./chat";
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 //@ts-ignore
 BigInt.prototype.toJSON = function () {
     return this.toString();
-};
-
-const defaultDirectChat: DirectChatSummary = {
-    kind: "direct_chat",
-    them: "a",
-    chatId: "abc",
-    readByMeUpTo: undefined,
-    readByThemUpTo: undefined,
-    latestMessage: {
-        event: {
-            kind: "message",
-            sender: "abcdefg",
-            messageId: newMessageId(),
-            messageIndex: 100,
-            content: {
-                kind: "text_content",
-                text: "some message",
-            },
-            reactions: [],
-            edited: false,
-            forwarded: false,
-        },
-        timestamp: BigInt(0),
-        index: 0,
-    },
-    latestEventIndex: 0,
-    dateCreated: BigInt(0),
-    notificationsMuted: false,
-    metrics: emptyChatMetrics(),
-    myMetrics: emptyChatMetrics(),
-    archived: false,
 };
 
 const defaultGroupChat: GroupChatSummary = {
@@ -107,20 +69,6 @@ const defaultGroupChat: GroupChatSummary = {
     archived: false,
     previewed: false,
 };
-
-function directChatId(id: number): DirectChatSummary {
-    return {
-        ...defaultDirectChat,
-        chatId: String(id),
-    };
-}
-
-function groupChatId(id: number): GroupChatSummary {
-    return {
-        ...defaultGroupChat,
-        chatId: String(id),
-    };
-}
 
 function createUser(userId: string, username: string, seconds: number): PartialUserSummary {
     const now = Date.now();
@@ -362,12 +310,6 @@ describe("updating poll votes", () => {
     });
 });
 
-describe("enough visible messages", () => {
-    test("returns false when there are no messages", () => {
-        expect(enoughVisibleMessages(true, [0, 1000], [])).toBe(false);
-    });
-});
-
 describe("get members string for group chat", () => {
     const withFewerThanSix = ["a", "b", "c", "d", "z"];
     const withUnknown = ["a", "b", "x", "d", "z"];
@@ -397,173 +339,4 @@ describe("get members string for group chat", () => {
         const members = getMembersString(user, lookup, withMoreThanSix, "Unknown User", "You");
         expect(members).toEqual("8 members");
     });
-});
-
-describe("merging updates", () => {
-    const initialChats = [
-        groupChatId(1),
-        groupChatId(2),
-        groupChatId(3),
-        directChatId(4),
-        directChatId(5),
-    ];
-
-    test("removed chats get removed", () => {
-        const updatesResponse: UpdatesResponse = {
-            chatsUpdated: [],
-            chatsRemoved: new Set(["1", "3", "5"]),
-            chatsAdded: [],
-            avatarIdUpdate: undefined,
-            timestamp: BigInt(0),
-            blockedUsers: undefined,
-            pinnedChats: undefined,
-            transactions: [],
-        };
-        const merged = mergeChatUpdates(initialChats, updatesResponse);
-        expect(merged.length).toEqual(2);
-        expect(merged[0].chatId).toEqual("2");
-        expect(merged[1].chatId).toEqual("4");
-    });
-
-    test("added chats get added", () => {
-        const updatesResponse: UpdatesResponse = {
-            chatsUpdated: [],
-            chatsRemoved: new Set([]),
-            chatsAdded: [directChatId(6), directChatId(7)],
-            avatarIdUpdate: undefined,
-            timestamp: BigInt(0),
-            blockedUsers: undefined,
-            pinnedChats: undefined,
-            transactions: [],
-        };
-        const merged = mergeChatUpdates(initialChats, updatesResponse);
-        expect(merged.length).toEqual(7);
-        expect(merged[6].chatId).toEqual("7");
-    });
-
-    describe("updated chats get merged correctly", () => {
-        const updatedDirect: DirectChatSummaryUpdates = {
-            kind: "direct_chat",
-            readByMeUpTo: 100,
-            chatId: "4",
-            readByThemUpTo: 200,
-            latestEventIndex: 300,
-            latestMessage: {
-                event: {
-                    kind: "message",
-                    content: {
-                        kind: "text_content",
-                        text: "test message",
-                    },
-                    sender: "abcdefg",
-                    repliesTo: undefined,
-                    messageId: newMessageId(),
-                    messageIndex: 300,
-                    reactions: [],
-                    edited: false,
-                    forwarded: false,
-                },
-                index: 300,
-                timestamp: BigInt(400),
-            },
-            affectedEvents: [],
-            metrics: emptyChatMetrics(),
-            myMetrics: emptyChatMetrics(),
-        };
-
-        const updatedGroup: GroupChatSummaryUpdates = {
-            kind: "group_chat",
-            chatId: "2",
-            lastUpdated: BigInt(1000),
-            readByMeUpTo: 250,
-            latestMessage: {
-                event: {
-                    kind: "message",
-                    content: {
-                        kind: "text_content",
-                        text: "test message",
-                    },
-                    sender: "2",
-                    repliesTo: undefined,
-                    messageId: newMessageId(),
-                    messageIndex: 300,
-                    reactions: [],
-                    edited: false,
-                    forwarded: false,
-                },
-                index: 300,
-                timestamp: BigInt(400),
-            },
-            latestEventIndex: 300,
-            name: "stuff",
-            description: "stuff",
-            mentions: [],
-            affectedEvents: [],
-            metrics: emptyChatMetrics(),
-            myMetrics: emptyChatMetrics(),
-            subtype: { kind: "no_change" },
-        };
-
-        test("attempting to update with a mismatched kind throws error", () => {
-            const updatesResponse: UpdatesResponse = {
-                chatsUpdated: [{ ...updatedDirect, chatId: "1" }],
-                chatsRemoved: new Set([]),
-                chatsAdded: [],
-                avatarIdUpdate: undefined,
-                timestamp: BigInt(0),
-                blockedUsers: undefined,
-                pinnedChats: undefined,
-                transactions: [],
-            };
-            expect(() => mergeChatUpdates(initialChats, updatesResponse)).toThrow();
-        });
-
-        test("direct chats get merged correctly", () => {
-            const updatesResponse: UpdatesResponse = {
-                chatsUpdated: [updatedDirect],
-                chatsRemoved: new Set([]),
-                chatsAdded: [],
-                avatarIdUpdate: undefined,
-                timestamp: BigInt(0),
-                blockedUsers: undefined,
-                pinnedChats: undefined,
-                transactions: [],
-            };
-            const merged = mergeChatUpdates(initialChats, updatesResponse);
-            const updated = merged.find((c) => c.chatId === "4");
-            if (updated && updated.kind === "direct_chat") {
-                expect(merged.length).toEqual(5);
-                expect(updated.readByMeUpTo).toEqual(100);
-                expect(updated.readByThemUpTo).toEqual(200);
-                expect(updated?.latestMessage).not.toBe(undefined);
-            } else {
-                fail("updated chat not found or was not a direct chat");
-            }
-        });
-
-        test("updated group chats get merged correctly", () => {
-            const updatesResponse: UpdatesResponse = {
-                chatsUpdated: [updatedGroup],
-                chatsRemoved: new Set([]),
-                chatsAdded: [],
-                avatarIdUpdate: undefined,
-                timestamp: BigInt(0),
-                blockedUsers: undefined,
-                pinnedChats: undefined,
-                transactions: [],
-            };
-            const merged = mergeChatUpdates(initialChats, updatesResponse);
-            const updated = merged.find((c) => c.chatId === "2");
-            if (updated && updated.kind === "group_chat") {
-                expect(merged.length).toEqual(5);
-                expect(updated.readByMeUpTo).toBe(250);
-                expect(updated?.lastUpdated).toEqual(BigInt(1000));
-                expect(updated?.latestMessage).not.toBe(undefined);
-            } else {
-                fail("updated chat not found or was not a group chat");
-            }
-        });
-    });
-
-    test.todo("chats end up in the right order");
 });
